# backend/openapi.yaml
openapi: '3.0.3'
info:
  title: "specimen-web-manage API"
  version: '0.0.2'
servers:
  - url: "http://localhost:8080"
    description: "local server"
paths:
  /create:
    get:
      tags:
        - "Occurrence"
      summary: "prepare create page"
      responses:
        '200':
          description: "get dropdown list and default value"
          content:
            application/json:
              schema:
                properties:
                  dropdown_list:
                    $ref: '#/components/schemas/Dropdown_list'
                  default_value:
                    $ref: '#/components/schemas/Occurrence'
              
    post:
      tags:
        - "Occurrence"
      summary: "create occurrence data"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              allOf:
                - $ref: '#/components/schemas/Occurrence'

      responses:
        "201":
            description: "Success Create occurrence data and return created occurrence data value"
            headers:
              Location:
                description: "URL of created resource"
                schema:
                  type: string
                  format: uri 
                  example: "/occurrence/1"
            content:
              application/json:
                schema:
                  $ref : '#/components/schemas/Occurrence'
        "400":
            description: "Bad Request"

        "500":
            description: "Server Error"
  /create/{occurrence_id}/attachments:
    post:
      tags:
        - "Occurrence"
      summary: "attach files to just created occurrence data"
      description: "attach files to just created occurrence data by occurrence_id if needed"
      parameters:
        - name: "occurrence_id"
          in: "path"
          description: "ID of occurrence to attach files"
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: false
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Attachment_files'

      responses:
        "200":
            description: "Success attach files to existing occurrence data"
            content:
              application/json:
                schema:
                  $ref : '#/components/schemas/Attachment_files'
        "400":
            description: "Bad Request"

        "500":
            description: "Server Error"
...
  /occurrences/{occurrence_id}/attachments:
      post:
        tags: ["Occurrence"]
        summary: "add new attachment files to existing occurrence data"
        parameters:
          - name: "occurrence_id"
            in: "path"
            required: true
            schema:
              type: integer
              example: 1
        requestBody:
          required: true
          content:
            multipart/form-data:
              schema:
                $ref: '#/components/schemas/Attachment_files'
        responses:
          '200':
            description: "Successfully added attachment files"
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Attachment_infos'
  /occurrences/{occurrence_id}/attachments/{attachment_id}:
    delete:
      tags: ["Occurrence"]
      summary: "delete attachments file from existing occurrence data"
      parameters:
        - name: "occurrence_id"
          in: "path"
          schema:
            type: integer
            example: 1
          required: true
      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Attachment_infos'
      responses:
        '204':
          description: "Successfully deleted attachment file"
...
components:
  schemas:
    User: 
      type: object
      properties:
        user_id: 
          type: integer
          example: 1
        user_name:
          type: string
          example: "admin"
    Users:
      type: array
      description: "list of users"
      items:
        $ref: '#/components/schemas/User'
      example:
        - user_id: 1
          user_name: "admin"
        - user_id: 2
          user_name: "user1"
...
    Project:
      type: object
      properties:
        project_id:
          type: integer
          example: 1
        project_name:
          type: string
          example: "test project"
    Projects:
      type: array
      description: "list of projects"
      items:
        $ref: '#/components/schemas/Project'
      example:
        - project_id: 1
          project_name: "test project"
        - project_id: 2
          project_name: "sample project"
...
    Language:
      type: object
      properties:
        language_id:
          type: integer
          example: 1
        language_common:
          type: string
          example: "English"
    Languages:
      type: array
      description: "list of languages"
      items:
        $ref: '#/components/schemas/Language'
      example:
        - language_id: 1
          language_common: "English"
        - language_id: 2
          language_common: "日本語" 
    Observation:
      type: object
      properties:
        observation_method_id:
          type: integer
          example: 1
        observation_method_name:
          type: string
          example: "Trap"
    Observations:
      type: array
      description: "list of observation methods"
      items:
        $ref: '#/components/schemas/Observation'
      example:
        - observation_method_id: 1
          observation_method_name: "Trap"
        - observation_method_id: 2
          observation_method_name: "Hand" 
...
    Specimen:
      type: object
      properties:
        specimen_methods_id:
          type: integer
          example: 1
        specimen_methods_common:
          type: string
          example: "Dried"
    Specimens:
      type: array
      description: "list of specimen methods"
      items:
        $ref: '#/components/schemas/Specimen'
      example:
        - specimen_methods_id: 1
          specimen_methods_common: "Dried"
        - specimen_methods_id: 2
          specimen_methods_common: "Preserved in alcohol"
...
    Institution:
      type: object
      properties:
        institution_id:
          type: integer
          example: 1
        institution_code:
          type: string
          example: "INST1"
    Institutions:
      type: array
      description: "list of institutions"
      items:
        $ref: '#/components/schemas/Institution'
      example:
        - institution_id: 1
          institution_code: "INST1"
        - institution_id: 2
          institution_code: "INST2"
...
    Attachment_file:
      properties:
        file:  
          type: string
          format: binary
          description: "file to be attached"
          example: "jpg file data"
        file_name:
          type: string
          description: "file name"
          example: "sample.jpg"
    Attachment_files:
      type: array
      items:
        $ref: '#/components/schemas/Attachment_file'
      description: "list of files to be attached"
      example: 
        - ["jpg file data","sample.jpg"]
        - ["wav file data","sample.wav"]
    Attachment_info:
      type: object
      properties:
        attachment_id:
          type: integer
          example: 1
        file_path:
          type: string
          example: "sample.jpg"
    Attachment_infos:
      type: array
      items:
        $ref: '#/components/schemas/Attachment_info'
      description: "list of attached file info"
      example:
        - attachment_id: 1
          file_path: "sample1.jpg"
        - attachment_id: 2
          file_path: "sample2.wav"
    Occurrence:
      type: object
      properties:
        user_id:
          type: integer
          example: 1
        project_id:
          type: integer
          example: 1
        individual_id:
          type: integer
          example: 1
        lifestage:
          type: string
          example: "adult" 
        sex:
          type: string 
          example: "male"
        body_lengh:
          type: string
          example: "10mm"
        creawted_at:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
        language_id:
          type: integer
          example: 1
# place 
        latigude:
          type: number
          format: float
          example: 35.6895
        longitude:
          type: number
          format: float
          example: 139.6917
#place_name is json format
        place_name:
          type: string
          example: "Tokyo shinjuku"

        note:
          type: string
          example: "sample note"

# cllasification 
        cllasification_id:
          type: integer
          example: 1
          readOnly: true
        species:
          type: string
          example: "fetida"
        genus:
          type: string
          example: "Eisenia"
        family:
          type: string
          example: "Lumbricidae"
        order:
          type: string
          example: "Haplotaxida"
        class:
          type: string
          example: "Clitellata"
        phylum:
          type: string
          example: "Annelida"
        kingdom:
          type: string
          example: "Animalia"
        others: # others contain verious sub-classificastion
          type: string
          example: "subspecies"
        
# observation
        observation:
          properties:
            observation_id:
              type: integer
              example: 1
              readOnly: true
            observation_user_id:
              type: integer
              example: 1 
            observation_method_id:
              type: integer
              example: 1 
            behavior:
              type: string
              example: "sample behavior"
            observed_at:
              type: string
              format: date-time
              example: "2024-01-01T00:00:00Z"
# specimen
        specimen:
          properties:
            specimen_id:
              type: integer
              example: 1
              readOnly: true
            specimen_user_id:
              type: integer
              example: 1
            specimen_methods_id:
              type: integer
              example: 1
            specimen_created_at:
              type: string
              format: date-time
              example: "2024-01-01T00:00:00Z"
            institution_id:
              type: integer
              example: 1
            collection_id:
              type: string
              example: "C-001"
# identification
        identification:
          properties:
            identification_id:
              type: integer
              example: 1
              readOnly: true
            identification_user_id:
              type: integer
              example: 1
            identified_at:
              type: string
              format: date-time
              example: "2024-01-01T00:00:00Z"
            source_info:
              type: string
              example: "sample identification sorce paper"
...
    Dropdown_list:
      type: object
      properties:
        users: 
          $ref: '#/components/schemas/Users'
        projects:
          $ref: '#/components/schemas/Projects'  
        languages:
          $ref: '#/components/schemas/Languages'
        observation_methods:
          $ref: '#/components/schemas/Observations'
        specimen_methods:
          $ref: '#/components/schemas/Specimens'
        institutions:
          $ref: '#/components/schemas/Institutions'
...